🚀 dots setup, project launch
Dotfiles are symlinked into the home directory and sourced in
`.bash_profile`.

The `Atomfile` contains a list of the Atom packages currently in
use. `atomwrite` will re-write the file, `atomload` will install
all packages from the file.

The `init.sh` script is meant to be run once, for initial setup.